[Sat] Data layer listener for non terrestrial networks
Adds a new field on MobileConnectionRepository, `isNonTerrestrial`,
which is a direct mapping from
ServiceState.isUsingNonterrestrialNetwork.

This is a property of the service state itself, and will be used to
determine the view state of the icon in future CLs. Due to there being
no listeners of this value yet, the code will not be called until the
(flagged) CL lands.

Test: MobileConnectionRepositoryTest
Test: DemoMobileConnectionRepository
Test: DemoMobileConnectionParameterizedTest
Bug: 311417356
Flag: NONE
Change-Id: If80e6f66f460f95bb225dccd88eae062cec621b6