Allow BAL permission for UserManagerService
On passing a pending intent to requestQuiteModeEnabled if authenticated
successfully then after disabling quiet mode it is used to start the
activity. This commit is to explicitly permit UserManagerService to
allow launch an activity from background by passing
ActivityOptions.MODE_BACKGROUND_ACTIVITY_START_ALLOWED when it calls
startIntentSender.
go/debug-bal

BUG: 319886822
Test: Build and verified manually

Change-Id: Ifeb887df410c851500df23ad9df85b2390200868