Get WSM per user service dynamically.
This change makes WearableSensingManagerService gets the current user's
per-user service instead of the default one. It also fixes an issue with
overriding the wearable sensing service implementation by shell command.
Previously, the WSS service name is cached in
WearableSensingManagerPerUserService and cannot be changed without a
phone restart, which means after the shell command overrides the service
value, it stays the same even if the command is run again with a different service name.
After this change, running the shell command will destroy
the cached per-user service and hence allows the service name to be
overwritten to a different value.

Bug: 301427767
Test: atest CtsWearableSensingServiceTestCases and then tried on-device
with a different WSS service name to verify the shell command can
overwrite the service name for a second time

Change-Id: Ie13b496f827e2f2e83da9d8a96ef9f85e6c93c8c