Use decor's TaskInfo instead of ShellTaskOrganizer's
When moving a task to front or doing an operation initiated from the
decoration, use the task info contained in the decoration instead of
trying to obtain it from ShellTaskOrganizer, which may produce a NPE.
This is a speculative fix to prevent the crash, but it should not be
possible for the decoration to exists and be handling input events if
the task has vanished and no longer cached by the organizer.

Bug: 319443130
Test: tap header to front, double-tap to resize and other header actions
work as expected.

Change-Id: Ia517ab47e79c5e80d9e6de3f310d8cc14b880649