Mark `HandlerThread` as daemon on Ravenwood.
Underlying test infrastructure waits for non-daemon threads to exit
at the end of any suite execution, which makes callers wait for the
unresponsive timeout to be triggered.

This change updates `HandlerThread` to be daemon so that we don't
cause the test suite execution to wait for any lingering threads
that a test author forgot to tear down.

Bug: 320520716
Test: atest FrameworksCoreTestsRavenwood:TestLooperManagerTest
Change-Id: I14a2f62dd67b059ed7ec19b13f9e16db8af68624