Escaping a curly brace to fix doc issue
The generated docs contained a pair of curly braces, like:

{{, 30}, {, 60}}

This is valid Javadoc and generated a valid HTMl output file, but the
file isn't publishable on DAC, because Devsite (our publishing engine)
interprets pairs of curly braces as a Jinja directive.

Fixed the issue by replacing one pair of braces with &lbrace; / &rbrace;

Built the docs with this change, and verified that the output can be
staged.

Change-Id: If579f03f4f99a9ba2e5219b98fa1eba4c4d47a1f
Test: go/abtd/run/L29600030001717486
Bug: 323000632