Updates event source to important parent only when events are merged
This is a partial revert of CL [1].

CL [1] changed two things:
- Makes sure that event source is an important node when sending a
  SUBTREE_CHANGED events. This event type is always sent to services,
  This ensures that accessibility services can recognize event sources.
- Adds SUBTREE_CHANGED type to the event change types when events from
  multiple nodes are merged into one.

The primary motivation of [1] was to make sure AccessibilityCache is
correctly invalidated, and this is accomplished only by the second
change. Further, although the first change aims to provide more correct
behavior, it resulted in a regression of some performance metrics.
Thus, this CL reverts the first part of change.

[1] Ic71f22ed6f6f54f49810dc83dc85dcd492dfc104

Bug: 322652084
Bug: 277305460
Test: ABTD to check the preformance regression.
Test: CtsAccessibilityTestCases CtsAccessibilityServiceTestCases
Change-Id: I1d9d4f6b2e4a1f34a97a3cc562f1a9c12ef67aaf