Move up showForced flag and background user verfication
This CL is a preparation for FocusController. It moves up the
showForced flag status verification and the validation to ensure
caller is not a background user. This refactoring is meant to
make the startInput logic move to FocusController easier.

This is just a refactoring CL, no behavior change is expected.

Bug: 314149476
Test: atest CtsInputMethodTestCases FrameworksServicesTests

Change-Id: Ib2e0c136c0a1a2794bac00723062699eff8f144a