Update minimum time window policy.
The new policy rejects jobs with short time windows and functional
constraints (eg. charging, connectivity, etc.). Jobs without functional
constraints and short time windows are allowed, but discouraged in a log
message. The minimum accepted window is also brought down to be in line
with the minimum job period.

The setOverrideDeadline documentation is also updated to make it clear
that the deadline just invalidates other requested (ie. functional)
constraints, but the job may be delayed by factors such as system health
signals.

Bug: 311402873
Test: atest CtsJobSchedulerTestCases:JobInfoTest
Test: atest frameworks/base/services/tests/mockingservicestests/src/com/android/server/job
Test: atest frameworks/base/services/tests/servicestests/src/com/android/server/job
Change-Id: I05def8a34e102c2e9b843d03f303f377fe4d4356