Delegation flow using connectionless handwriting
The handwriting delegation flow is updated to use a connectionless
handwriting session when the IME supports it. When there is stylus
movement on a delegator view, a connectionless handwriting session will
be started. If it completes successfully and returns a result, the
delegation transition will be triggered, and the delegate editor view
will receive the recognized text when it creates an input connection.

When the IME does not support connectionless handwriting session, the
old handwriting delegation flow will be used.

Bug: 300979854
Test: atest StylusHandwritingTest
Change-Id: I1bd53784a4005cf992d74adb03bb27dd4fcc6d8e